git-parser: fix color reset for completion bug
The colors were not properly reset, which lead to the completion not
properly clearing the line. See https://superuser.com/a/893846 for a
description.

The reason is that the length of the prompt is calculated containing the
color escape codes, but they are obviously interpreted by the terminal.
Adding %{ %} around parts tells zsh that these characters don't count
for the length of the prompt